TG Initial rebase

+++ code/game/objects/items/grenades/antigravity.dm
@@ -1,17 +1,21 @@
 /obj/item/grenade/antigravity
 	name = "antigravity grenade"
 	icon_state = "emp"
-	item_state = "emp"
-	
+	inhand_icon_state = "emp"
+
 	var/range = 7
 	var/forced_value = 0
 	var/duration = 300
 
-/obj/item/grenade/antigravity/prime()
+/obj/item/grenade/antigravity/detonate(mob/living/lanced_by)
+	. = ..()
+	if(!.)
+		return
+
 	update_mob()
-	
-	for(var/turf/T in view(range,src))
-		var/datum/component/C = T.AddComponent(/datum/component/forced_gravity,forced_value)
-		QDEL_IN(C,duration)
-		
+
+	for(var/turf/lanced_turf in view(range, src))
+		lanced_turf.AddElement(/datum/element/forced_gravity, forced_value)
+		addtimer(CALLBACK(lanced_turf, TYPE_PROC_REF(/datum/, _RemoveElement), list(/datum/element/forced_gravity, forced_value)), duration)
+
 	qdel(src)
